# perf(table) Coalesces style updates after style measurements to reduc…