--- a/src/cdk/table/public-api.ts
+++ b/src/cdk/table/public-api.ts
@@ -8,6 +8,7 @@
 
 export * from './table';
 export * from './cell';
+export * from './coalesced-style-scheduler';
 export * from './row';
 export * from './table-module';
 export * from './sticky-styler';