--- a/src/cdk/table/sticky-styler.ts
+++ b/src/cdk/table/sticky-styler.ts
@@ -156,7 +156,8 @@ export class StickyStyler {
       }
     }
 
-    // Coalesce with other sticky row updates (top/bottom), sticky columns updates (and potentially other changes like column resize).
+    // Coalesce with other sticky row updates (top/bottom), sticky columns updates
+    //(and potentially other changes like column resize).
     this._coalescedStyleScheduler.schedule(() => {
       for (let rowIndex = 0; rowIndex < rows.length; rowIndex++) {
         if (!states[rowIndex]) {