… src/cdk/table/table.spec.ts
… src/material-experimental/column-resize/column-resize.spec.ts
… src/material-experimental/mdc-table/table.spec.ts
… src/material/table/table.spec.ts
… tools/public_api_guard/cdk/table.d.ts