…e layout thrashing Exposes the CoalescedStyleScheduler for use by other related components in a table such as column resize.

This will complement #19739, but the following measurements are taken without it. Each of these reduces the impact of the other but they'll both result in some savings (especially this one). I'm exposing the coalesced style scheduler so that it can also be used by column resize et al to coalesce along with sticky styler.

Before:
![image](https://user-images.githubusercontent.com/923677/85632044-94480280-b644-11ea-8b50-c7ea8eda9d4c.png)
94ms script + 46ms rendering

After:
![image](https://user-images.githubusercontent.com/923677/85632118-b2adfe00-b644-11ea-8c77-a610dbce9b24.png)
69ms script + 22ms rendering
This is scrolled down a little too far to see the effect - we're now down to one compute styles + layout before data rows are rendered and one after. Previously there were many.